Add support for advanced blink-reconstruction mode - Requires datamatrix >= 0.13.0 - Bump to 0.12.0

+++ eyelinkparser/_eyelinkparser.py
@@ -356,7 +356,7 @@ class EyeLinkParser(object):
                 (None, u'fixstlist_', self.fixstlist),
                 (None, u'fixetlist_', self.fixetlist),
                 ]):
-            trace = np.array(trace)
+            trace = np.array(trace, dtype=float)
             if tracelabel is not None and self._traceprocessor is not None:
                 trace = self._traceprocessor(tracelabel, trace)
             if self._maxtracelen is not None \
